apis/resume-jobdescription-scorer/requirements.txt
spacy
flask
fastapi
uvicorn
python-dotenv